Refactor hello test controller package
