+++ be/src/main/java/com/app/be/hello/Hello.java
@@ -1,4 +1,4 @@
-package com.app.be;
+package com.app.be.hello;
 
 public record Hello(long id, String content) {
 }
